# MS-deal/src/main/resources/openapi/deal-api.yaml
openapi: 3.1.0
info:
  title: Deal Service API
  version: 0.8.0
  description: Микросервис оформления заявок на кредит.

servers:
  - url: http://localhost:9020

paths:
  /deal/statement:
    post:
      summary: Создать заявку и получить предложения по кредиту
      operationId: createLoanStatement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanStatementRequestDto'
      responses:
        '200':
          description: Список предложений по кредиту
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanOfferDto'
        '400':
          description: Неверный запрос
        '500':
          description: Внутренняя ошибка сервера
  /deal/offer/select:
    post:
      summary: Выбрать интересующее предложение по кредиту
      operationId: selectLoanOffer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanOfferDto'
      responses:
        '200':
          description: Успешный выбор предложения
        '404':
          description: Заявка не найдена
  /deal/calculate/{statementId}:
    post:
      summary: Рассчитать итоговые параметры кредита
      operationId: calculateFinalLoanParameters
      parameters:
        - name: statementId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinishRegistrationRequestDto'
      responses:
        '200':
          description: Успешный расчет итоговых параметров кредита'
        '404':
          description: Заявка не найдена
        '500':
          description: Внутренняя ошибка сервера


components:
  schemas:
    LoanStatementRequestDto:
      type: object
      properties:
        amount:
          type: number
          format: double
          description: Сумма кредита
          example: 500000.0
        term:
          type: integer
          description: Срок кредита в месяцах
          example: 12
        firstName:
          type: string
          description: Имя клиента
          example: John
        lastName:
          type: string
          description: Фамилия клиента
          example: Smith
        middleName:
          type: string
          description: Отчество клиента
          example: Joseph
        email:
          type: string
          description: Электронная почта клиента
          example: example@neoflex.ru
        birthdate:
          type: string
          format: date
          description: Дата рождения клиента
          example: "1990-01-01"
        passportSeries:
          type: string
          description: Серия паспорта клиента
          example: "1234"
        passportNumber:
          type: string
          description: Номер паспорта клиента
          example: "567890"
      required: [ amount, term, firstName, lastName, email, birthdate, passportSeries, passportNumber ]

    LoanOfferDto:
      type: object
      properties:
        statementId:
          type: string
          format: uuid
          description: ID заявки
          example: "5d5c2f1a-2e23-4dc7-9c5c-b18b1a4c7a77"
        requestedAmount:
          type: number
          format: double
          description: Запрашиваемая сумма кредита
          example: 500000.0
        totalAmount:
          type: number
          format: double
          description: Итоговая сумма кредита с учетом страховки
          example: 550000.0
        term:
          type: integer
          description: Срок кредита в месяцах
          example: 12
        monthlyPayment:
          type: number
          format: double
          description: Ежемесячный платеж по кредиту
          example: 45000.0
        rate:
          type: number
          format: double
          description: Процентная ставка по кредиту
          example: 14.0
        isInsuranceRequired:
          type: boolean
          description: Наличие опции страхования
          example: true
        isSalaryClient:
          type: boolean
          description: Является ли клиент зарплатным
          example: true
      required: [ statementId, requestedAmount, totalAmount, term, monthlyPayment, rate, isInsuranceRequired, isSalaryClient ]

    FinishRegistrationRequestDto:
      type: object
      properties:
        gender:
          $ref: '#/components/schemas/Gender'
        maritalStatus:
          $ref: '#/components/schemas/MaritalStatus'
        dependentAmount:
          type: integer
        passportIssuedDate:
          type: string
          format: date
        passportIssueBranch:
          type: string
        employment:
          $ref: '#/components/schemas/EmploymentDto'
        accountNumber:
          type: string
      required: [ gender, maritalStatus, dependentAmount, passportIssuedDate, passportIssueBranch, employment, accountNumber ]
      example:
        gender: MALE
        maritalStatus: SINGLE
        dependentAmount: 0
        passportIssuedDate: 2010-01-01
        passportIssueBranch: "770-001"
        employment:
          employmentStatus: EMPLOYED
          employerINN: "1234567890"
          salary: 80000.0
          position: MID_MANAGER
          workExperienceTotal: 120
          workExperienceCurrent: 36
        accountNumber: "40817810099910004312"


    EmploymentDto:
      type: object
      properties:
        employmentStatus:
          $ref: '#/components/schemas/EmploymentStatus'
        employerINN:
          type: string
          description: ИНН работодателя
          example: "1234567890"
        salary:
          type: number
          format: double
          description: Зарплата клиента
          example: 80000.0
        position:
          $ref: '#/components/schemas/Position'
        workExperienceTotal:
          type: integer
          description: Общий стаж работы (в месяцах)
          example: 120
        workExperienceCurrent:
          type: integer
          description: Стаж работы на текущем месте (в месяцах)
          example: 10
      required:
        - employmentStatus
        - employerINN
        - salary
        - position
        - workExperienceTotal
        - workExperienceCurrent

    Gender:
      type: string
      enum: [MALE, FEMALE, NON_BINARY]

    MaritalStatus:
      type: string
      enum: [MARRIED, SINGLE, DIVORCED, WIDOW_WIDOWER]

    EmploymentStatus:
      type: string
      enum: [ EMPLOYED, UNEMPLOYED, SELF_EMPLOYED, BUSINESS_OWNER ]

    Position:
      type: string
      enum: [ WORKER, MID_MANAGER, TOP_MANAGER, OWNER ]